# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/shows/search": {
      "get": {
        "operationId": "ShowsController_searchShows",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchShowsResultsDto"
                }
              }
            }
          }
        },
        "summary": "Search shows from TMDB",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/discover": {
      "get": {
        "operationId": "ShowsController_discoverShows",
        "parameters": [
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort order for results",
            "schema": {
              "default": "popularity.desc",
              "type": "string",
              "enum": [
                "popularity.desc",
                "popularity.asc",
                "first_air_date.desc",
                "first_air_date.asc",
                "vote_average.desc",
                "vote_average.asc"
              ]
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "description": "Filter by first air date year",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchShowsResultsDto"
                }
              }
            }
          }
        },
        "summary": "Discover popular shows from TMDB",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/tmdb/{showId}": {
      "get": {
        "operationId": "ShowsController_getShowDetails",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TMDBShowDetailDto"
                }
              }
            }
          }
        },
        "summary": "Get show details from TMDB",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/tmdb/{showId}/credits": {
      "get": {
        "operationId": "ShowsController_getFullShowCredits",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullCreditsDto"
                }
              }
            }
          }
        },
        "summary": "Get the full cast and crew for a show, crew grouped by department",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/tmdb/{showId}/watch-providers": {
      "get": {
        "operationId": "ShowsController_getWatchProviders",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "ISO 3166-1 country code (e.g. US, GB)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchProvidersResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get watch providers for a show from TMDB/JustWatch",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/tmdb/{showId}/recommendations": {
      "get": {
        "operationId": "ShowsController_getRecommendations",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchShowsResultsDto"
                }
              }
            }
          }
        },
        "summary": "Get TMDB recommendations (similar shows)",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/tmdb/{showId}/season/{seasonNumber}": {
      "get": {
        "operationId": "ShowsController_getSeasonDetails",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seasonNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TMDBSeasonDetailDto"
                }
              }
            }
          }
        },
        "summary": "Get season details from TMDB",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/tmdb/{showId}/local-seasons": {
      "get": {
        "operationId": "ShowsController_getLocalSeasons",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalSeasonDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get seasons from local database (populated via sync on first interaction)",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/tmdb/{showId}/season/{seasonNumber}/local-episodes": {
      "get": {
        "operationId": "ShowsController_getLocalEpisodes",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seasonNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalEpisodeDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get episodes from local database (populated via sync on first interaction)",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/tmdb/{showId}/season/{seasonNumber}/episode/{episodeNumber}": {
      "get": {
        "operationId": "ShowsController_getEpisodeDetails",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seasonNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "episodeNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TMDBEpisodeDto"
                }
              }
            }
          }
        },
        "summary": "Get episode details from TMDB",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/user/{userDid}": {
      "get": {
        "operationId": "ShowsController_getUserShows",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackedShowSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get tracked shows for a user",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/user/{userDid}/up-next": {
      "get": {
        "operationId": "ShowsController_getUserUpNext",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to return (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return per page",
            "schema": {
              "default": 8,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "default": "lastWatched",
              "type": "string",
              "enum": [
                "lastWatched",
                "title",
                "progress"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort direction",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "showId",
            "required": false,
            "in": "query",
            "description": "Only return up next for this show",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUpNextResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get up next episodes for a user",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/user/{userDid}/release-calendar": {
      "get": {
        "operationId": "ShowsController_getUserReleaseCalendar",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for release calendar range (YYYY-MM-DD)",
            "schema": {
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for release calendar range (YYYY-MM-DD)",
            "schema": {
              "example": "2026-03-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseCalendarResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get releases for watched shows and watchlist items within a date range",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/watched": {
      "post": {
        "operationId": "ShowsController_markWatched",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkEpisodeWatchedDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedEpisodeDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Mark an episode as watched",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/watched/{showId}": {
      "delete": {
        "operationId": "ShowsController_unmarkWatched",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "description": "Remove mode: latest (default) removes most recent, all removes all",
            "schema": {
              "enum": [
                "latest",
                "all"
              ],
              "type": "string"
            }
          },
          {
            "name": "episodeNumber",
            "required": false,
            "in": "query",
            "description": "Optional filter by episode",
            "schema": {}
          },
          {
            "name": "seasonNumber",
            "required": false,
            "in": "query",
            "description": "Optional filter by season",
            "schema": {}
          }
        ],
        "responses": {
          "204": {
            "description": "Episode unmarked as watched"
          }
        },
        "summary": "Unmark episode(s) as watched",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/{showId}": {
      "get": {
        "operationId": "ShowsController_getShow",
        "parameters": [
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedShowSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Get show from database",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/user/{userDid}/show/{showId}/history": {
      "get": {
        "operationId": "ShowsController_getShowWatchHistory",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "description": "User DID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "showId",
            "required": true,
            "in": "path",
            "description": "TMDB show ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Watch history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EpisodeHistoryItemDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Get watch history for a specific show",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/progress": {
      "post": {
        "operationId": "ShowsController_getShowProgress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShowProgressBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowProgressBatchResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get viewer progress for a batch of shows",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/history/{trackedEpisodeId}": {
      "delete": {
        "operationId": "ShowsController_deleteEpisodeWatchHistoryEntry",
        "parameters": [
          {
            "name": "trackedEpisodeId",
            "required": true,
            "in": "path",
            "description": "Tracked episode entry ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Watch history entry deleted"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Tracked episode entry not found"
          }
        },
        "summary": "Delete a specific episode watch history entry",
        "tags": [
          "shows"
        ]
      }
    },
    "/shows/season/watched": {
      "post": {
        "operationId": "ShowsController_markSeasonWatched",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkSeasonWatchedDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkedEpisodesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Mark all episodes in a season as watched",
        "tags": [
          "shows"
        ]
      }
    },
...
  },
  "info": {
    "title": "Opnshelf API",
    "description": "Personal media tracker powered by AT Protocol",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "SearchShowsResultsDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total count of items across all pages"
          },
          "page": {
            "type": "number",
            "description": "Current page number after server-side clamping"
          },
          "pageSize": {
            "type": "number",
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of available pages"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether a next page exists"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Whether a previous page exists"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TMDBShowResultDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage",
          "items"
        ]
      },
...
    }
  }
}